+++ crates/ruma-client-api/src/push.rs
@@ -286,9 +286,13 @@ impl PusherIds {
 }
 
 /// Information for an email pusher.
-#[derive(Clone, Debug, Default)]
+#[derive(Clone, Debug, Default, Serialize, Deserialize)]
 #[cfg_attr(not(feature = "unstable-exhaustive-types"), non_exhaustive)]
-pub struct EmailPusherData;
+#[serde(transparent, default)]
+pub struct EmailPusherData {
+    /// Custom data for the pusher.
+    pub data: JsonObject,
+}
 
 impl EmailPusherData {
     /// Creates a new empty `EmailPusherData`.