+++ crates/ruma-client-api/CHANGELOG.md
@@ -3,6 +3,8 @@
 Improvements:
 
 - Add unstable support for reporting rooms, according to MSC4151.
+- `EmailPusherData` allows to set custom data for the pusher in the `data` field, due
+  to a clarification in the spec.
 
 # 0.19.0